build: solution to the pnpm linking issue in the playground
When using a non-workspace setup with pnpm link in the playground as is the
default in the module starter, deps (like the @witchraft/ui dep) can resolve to
different instances, causing all sorts of issues with injected key errors (since
the symbols are no longer equal).

Switched to pnpm workspace + catalog which also helps keep deps the same
version. We do sacrifice being able to set the version of module deps from the
package json in the nuxt module.